+++ packages/integration-kit/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @trigger.dev/integration-kit
 
+## 2.0.5
+
+## 2.0.4
+
 ## 2.0.3
 
 ## 2.0.2