+++ packages/react/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @trigger.dev/react
 
+## 2.0.5
+
+### Patch Changes
+
+- @trigger.dev/core@2.0.5
+
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies [96384991]
+  - @trigger.dev/core@2.0.4
+
 ## 2.0.3
 
 ### Patch Changes